sdk: Add Makefile and Rust app examples
