+++ sdk/Cargo.toml
@@ -1,11 +1,11 @@
 [package]
-name = "islet-sdk"
+name = "islet_sdk"
 version = "0.1.0"
 authors = ["Islet Contributors"]
 edition = "2021"
 
 [lib]
-crate-type = ["cdylib"]
+crate-type = ["rlib", "cdylib"]
 
 [build-dependencies]
 cbindgen = "0.24.0"
